Updated the Scene's atmosphere heavily.
Added fog texture and applied it to the Borders.
Added Harpoons, for some nice environmental assets.
Carved out the Yellow Key Door behind the Shrine (Currently works with the Red Key)
Made the bridges near the spawn room into actual bridges by pulling up the bottom centre of them, making them look a lot nicer.
Redistributed the Sea of Hands to the entirety of the spawn area, recoloured and resized them.
Added a skybox, and modified it's texture to 128x Pixels.
Added a Sea texture, and added it to the scene. Modified it.
Modified the post-drop drown zone, moved one of the lanterns around and pushed it all back a bit to make room for the opposite exit.